Project commit 3

+++ src/main/java/com/example/service/MessageService.java
@@ -1,6 +1,7 @@
 package com.example.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -29,9 +30,27 @@ public class MessageService {
         return messageRepository.findAll();
    }
 
-   public Message deleteById(int id){
-        Message deleted = messageRepository.getById(id);
+   public Message deleteById(int id) {
+    Optional<Message> deleted = messageRepository.findById(id);
+    if (deleted.isPresent()) {
         messageRepository.deleteById(id);
-        return deleted;
-   }
+        return deleted.get();
+    }
+    return null;
+    }
+
+    public Message getMessageById(int id) {
+        Optional<Message> message = messageRepository.findById(id);
+        if (message.isPresent()) {
+            return message.get();
+        }
+        return null;
+    }
+
+    public List<Message> getAllMessageByUser(int accountId){
+        List<Message> messages = messageRepository.getMessageByPostedBy(accountId);
+        return messages;
+    }
+
+
 }
